+++ CHANGELOG.md
@@ -18,8 +18,9 @@
     - use options instead of arguments
     - allow to run for multiple databases (or for all databases)
 - Command `odood db list-installed-addons` renamed to `odood addons find-installed`.
-- Command `odood addons find-installed` got one more option:
+- Command `odood addons find-installed` got new options:
     - `--non-system` - output only non-system addons (that are not included in official Odoo community)
+    - `--format` - what format to use for output: list, assembly-spec
 
 ---